move more runloop functiosn to runloop.c
