+++ retroarch.h
@@ -226,6 +226,14 @@ enum runloop_action
    RUNLOOP_ACTION_AUTOSAVE
 };
 
+enum input_game_focus_cmd_type
+{
+   GAME_FOCUS_CMD_OFF = 0,
+   GAME_FOCUS_CMD_ON,
+   GAME_FOCUS_CMD_TOGGLE,
+   GAME_FOCUS_CMD_REAPPLY
+};
+
 #ifdef HAVE_BSV_MOVIE
 enum rarch_movie_type
 {
@@ -2391,6 +2399,22 @@ bool bsv_movie_check(
 void bsv_movie_deinit(runloop_state_t *p_runloop);
 bool bsv_movie_init(runloop_state_t *p_runloop);
 #endif
+
+void runloop_system_info_free(runloop_state_t *p_runloop);
+void runloop_fastmotion_override_free(settings_t *settings,
+      runloop_state_t *p_runloop);
+void runloop_game_focus_free(runloop_state_t *p_runloop);
+void runloop_audio_buffer_status_free(runloop_state_t *p_runloop);
+void runloop_frame_time_free(runloop_state_t *p_runloop);
+#ifdef HAVE_RUNAHEAD
+void runloop_runahead_clear_variables(runloop_state_t *p_runloop);
+#endif
+retro_time_t runloop_set_frame_limit(
+      const struct retro_system_av_info *av_info,
+      float fastforward_ratio);
+float runloop_get_fastforward_ratio(
+      settings_t *settings,
+      struct retro_fastforwarding_override *fastmotion_override);
 
 RETRO_END_DECLS
